Split native audio options DTO module
